Refactored bird detail button view
